emu-core: Widen Cpu::pc() to u32, add WordBus trait
Phase 1 of the implementation roadmap.

Cpu::pc() now returns u32 to support all CPU address widths (16-bit 6502/Z80,
24-bit 68000, 32-bit ARM7TDMI). The 6502 and Z80 implementations zero-extend
their 16-bit PC via u32::from().

WordBus extends Bus with read_word/write_word for 16-bit data bus systems
like the 68000 and SH-2. These CPUs also do byte accesses (e.g. CIA
registers), so WordBus is a supertrait of Bus rather than a replacement.

Clocked trait (from the plan) already exists as Tickable — no action needed.

Entire-Checkpoint: 2cf7335e2fb1

+++ crates/emu-core/src/cpu.rs
@@ -21,7 +21,10 @@ pub trait Cpu {
     fn tick<B: Bus>(&mut self, bus: &mut B);
 
     /// Returns the current program counter.
-    fn pc(&self) -> u16;
+    ///
+    /// Returns `u32` to support all CPU address widths: 16-bit (6502, Z80),
+    /// 24-bit (68000), and 32-bit (ARM7TDMI). Narrower CPUs zero-extend.
+    fn pc(&self) -> u32;
 
     /// Returns a snapshot of all registers for inspection.
     fn registers(&self) -> Self::Registers;
